requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
kombu==4.5.0
Markdown==3.1
mysqlclient==1.4.2.post1
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
vine==1.3.0
pylint==2.3.1
flower==0.9.3
django-render-block==0.6
requests==2.22.0
sentry-sdk==0.8.0